+++ workspaces/kiali/plugins/kiali/src/pages/WorkloadDetails/WorkloadsDescription.tsx
@@ -74,7 +74,7 @@ export const infoStyle = kialiStyle({
 
 export const healthIconStyle = kialiStyle({
   marginLeft: '0.5rem',
-  verticalAlign: '-0.075rem',
+  verticalAlign: '+0.275rem',
 });
 
 const additionalItemStyle = kialiStyle({